pyproject.toml
[project]
name = "cyteonto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "anndata>=0.12.2",
    "asyncio>=4.0.0",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "logfire>=4.3.3",
    "loguru>=0.7.3",
    "mypy>=1.17.1",
    "numpy>=2.3.2",
    "owlready2>=0.48",
    "pandas>=2.3.1",
    "pydantic-ai>=0.7.4",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.9",
    "scikit-learn>=1.7.1",
    "tenacity>=9.1.2",
    "tomli>=2.2.1",
    "tqdm>=4.67.1",
    "xmltodict>=0.14.2",
]
